% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_of_birth.R
\name{date_of_birth}
\alias{date_of_birth}
\title{A tools4ukbb function
Function output: a column of name "DOB" (in yyyy-mm-dd date format), column of eid values}
\usage{
date_of_birth(dataframe)
}
\arguments{
\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tools4ukbb function
Function output: a column of name "DOB" (in yyyy-mm-dd date format), column of eid values
}
\examples{
date_of_birth()
}
\keyword{birth}
\keyword{date}
\keyword{of}
