% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_date.R
\name{dx_date}
\alias{dx_date}
\title{A tools4ukbb function}
\usage{
dx_date(icd_list, dataframe, disease_name)
}
\arguments{
\item{icd_list}{a list of the icd10 codes you wish to investigate}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (502462 x 18158 as of 09/07/2021)}

\item{disease_name}{a string containing the name of the disease(s) of interest}
}
\description{
Function output: a column of name "Date_of_first_{disease_name}\emph{diagnosis", column of eid values
"Date_of_first}{disease_name}_diagnosis" contains the date an individual received the diagnosis (as indicated by the icd_code(s) of interest)
}
\examples{
dx_date()
}
\keyword{date}
