% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_hx.R
\name{dx_hx}
\alias{dx_hx}
\title{A tools4ukbb function}
\usage{
dx_hx(icd_list, dataframe, disease_name)
}
\arguments{
\item{icd_list}{a list of the icd10 codes you wish to investigate}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}

\item{disease_name}{a string containing the name of the disease(s) of interest}
}
\description{
Function output: a column of name "Hx_of_{disease_name}", column of eid values
"Hx_of_{disease_name}" contains 0's (indicating an absence of the diagnosis) or 1's (indicating the presence of a diagnosis)
}
\examples{
dx_hx()
}
\keyword{hx}
