% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_with_disease.R
\name{individuals_with_disease}
\alias{individuals_with_disease}
\title{A tools4ukbb function
Function output: a list of all individuals (identified by eid) that have been diagnosed with the icd_codes of interest}
\usage{
individuals_with_disease(icd_list, dataframe)
}
\arguments{
\item{icd_list}{a list of the icd10 codes you wish to investigate}

\item{dataframe}{the originial phenotype dataframe containing all individuals in the ukbiobank (~500,000 cols x 18,000 rows as of 09/07/2021)}
}
\description{
A tools4ukbb function
Function output: a list of all individuals (identified by eid) that have been diagnosed with the icd_codes of interest
}
\examples{
individuals_with_disease()
}
\keyword{with}
